requirements.txt
Flask==3.0.2
gunicorn==21.2.0
python-dotenv==1.0.1
requests==2.31.0
redis==5.0.1
rq==1.15.1
pydantic==2.6.3
pydantic-settings==2.1.0
